tests: Split the tests and helper functions into separate files
