+++ Makefile.am.service.inc
@@ -222,6 +222,7 @@ test_data = \
 test_support = \
 	test/__init__.py \
 	test/build_app.py \
+	test/service_test_helpers.py \
 	$(NULL)
 
 TESTS_ENVIRONMENT = \
