+++ com.endlessm.CompanionAppService.PipDependencies.json
